fix NavBar new component
