+++ client/package.json
@@ -16,6 +16,7 @@
     "react-redux": "^7.2.3",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.1",
+    "redux-thunk": "^2.4.2",
     "styled-components": "^6.0.7",
     "web-vitals": "^2.1.4"
   },
