+++ client/src/App.js
@@ -1,8 +1,6 @@
 import './App.css';
 
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';        
-import NavBar from './components/NavBar/NavBar';
-import Footer from './components/Footer/Footer';
 import Landing from './View/Landing/Landing';
 import Home from './components/Home/Home'
 import ProductsDetail from '../src/View/Detail/Detail'
@@ -13,16 +11,11 @@ function App() {
   return (
     <div className="App">
 
-      {/* <NavBar/>
-      <Landing/>
-      <Home/>
-      <Footer/> */}
       <Router>
       <Switch>
         <Route path="/" exact component={Landing} />
         <Route path="/home" component={Home} />
-        <Route path="/detail/:id" component={ProductsDetail} />
-       <Route path="/detail/:id"/>
+        <Route path="/product/:id" component={ProductsDetail} />
       </Switch>
     </Router>
     </div>
