+++ client/src/View/Landing/Landing.jsx
@@ -1,49 +1,15 @@
-import { Container, Nav, Navbar, NavDropdown, Col, Image, Row, Modal } from 'react-bootstrap' 
+import { Container, Nav, Col, Image, Row } from 'react-bootstrap' 
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './Landing.css';
-import { Link } from 'react-router-dom';
-import { BsCart3 } from 'react-icons/bs'
-import { FiLogIn } from 'react-icons/fi'
-import { useState } from 'react';
+import  NavBar  from '../../components/NavBar/NavBar';
+import { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
 
 function Landing(){
-
-    const [smShow, setSmShow] = useState(false);
+    
     return(
     <div>
-      <Navbar bg="dark" expand="lg" variant="dark">
-        <Container>
-          <img src="logo.png" alt="Logo" height="45px" className='custom-nav'/>       
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto custom-nav">
-            <Nav.Link as={Link} to="/home">Home</Nav.Link>
-
-              <Navbar.Brand href="#home">
-                  <BsCart3 className="nav-icon me-2" onClick={() => setSmShow(true)} />
-                  <Modal
-                    size="sm"
-                    show={smShow}
-                    onHide={() => setSmShow(false)}
-                    aria-labelledby="example-modal-sizes-title-sm"
-                  >
-                  <Modal.Header closeButton>
-                    <Modal.Title id="example-modal-sizes-title-sm">
-                      Carrito
-                    </Modal.Title>
-                  </Modal.Header>
-                  <Modal.Body>Registrate o inicia sesion para ver o modificar el carrito !</Modal.Body>
-                  </Modal>
-                </Navbar.Brand>
-
-              <Navbar.Brand>
-                <FiLogIn className="nav-icon"/>
-              </Navbar.Brand>
-
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
+      <NavBar/>
       <Container>
       <Row className='offer-image'>
         <Col xs={6} md={4} lg={4}>
