+++ client/src/components/Home/Home.jsx
@@ -2,9 +2,13 @@ import Nav from 'react-bootstrap/Nav';
 import styles from './Home.module.css'
 import CardProducto from '../CardProducto/CardProducto'
 import products from '../../data'
+import NavBar from '../NavBar/NavBar';
 
 export default function Home() {
+
     return (
+        <div className='Home'>
+        <NavBar/>
         <div className={styles.container}>
 
             <Nav className={styles.side_bar}>
@@ -36,5 +40,6 @@ export default function Home() {
             </div>
 
         </div >
+        </div>
     )
 }
